// amplify/functions/create-campaign/graphql/API.ts
export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}
// ...
export type GetAcquisitionQueryVariables = {
  acquisitionId: string,
};

export type GetAcquisitionQuery = {
  getAcquisition?:  {
    __typename: "Acquisition",
    acquisitionId: string,
    campaignId: string,
    campaignPost?:  {
      __typename: "CampaignPost",
      campaignId: string,
      campaignPostId: string,
      clientAccountId: string,
      country?: string | null,
      createdAt: string,
      createdBy: string,
      disabled?: boolean | null,
      email?: string | null,
      emailVerified?: string | null,
      familyName?: string | null,
      givenName?: string | null,
      language?: string | null,
      linkedInPostId?: string | null,
      name?: string | null,
      pictureUrl?: string | null,
      postType?: string | null,
      postTypeEnum?: CampaignPostPostTypeEnum | null,
      readGroups?: Array< string | null > | null,
      sub?: string | null,
      updatedAt: string,
      updatedBy: string,
      writeGroups?: Array< string | null > | null,
    } | null,
    createdAt: string,
    createdBy: string,
    disabled?: boolean | null,
    readGroups?: Array< string | null > | null,
    updatedAt: string,
    updatedBy: string,
    viewedCampaignPostId: string,
    writeGroups?: Array< string | null > | null,
  } | null,
};

export type GetCampaignQueryVariables = {
  campaignId: string,
};

export type GetCampaignQuery = {
  getCampaign?:  {
    __typename: "Campaign",
    campaignId: string,
    campaignPostClicks?:  {
      __typename: "ModelCampaignPostClickConnection",
      nextToken?: string | null,
    } | null,
    campaignPostStates?:  {
      __typename: "ModelCampaignPostStateConnection",
      nextToken?: string | null,
    } | null,
    campaignPosts?:  {
      __typename: "ModelCampaignPostConnection",
      nextToken?: string | null,
    } | null,
    clientAccountId: string,
    confirmationPageUrl?: string | null,
    createdAt: string,
    createdBy: string,
    description?: string | null,
    disabled?: boolean | null,
    event?:  {
      __typename: "Event",
      clientAccountId: string,
      closingDate: string,
      createdAt: string,
      createdBy: string,
      description?: string | null,
      disabled?: boolean | null,
      eventId: string,
      location?: string | null,
      name: string,
      openingDate: string,
      readGroups?: Array< string | null > | null,
      updatedAt: string,
      updatedBy: string,
      websiteUrl?: string | null,
      writeGroups?: Array< string | null > | null,
    } | null,
    eventId: string,
    modalIncentiveTCs?: string | null,
    modalIncentiveText?: string | null,
    modalPattern?: string | null,
    modalPatternLocationEnum?: CampaignModalPatternLocationEnum | null,
    modalText?: string | null,
    modalTitle?: string | null,
    name?: string | null,
    postForwardUrl?: string | null,
    postImageUrl?: string | null,
    postText?: string | null,
    readGroups?: Array< string | null > | null,
    registrationPageUrl?: string | null,
    targetInfluencerTypeEnum?: CampaignTargetInfluencerTypeEnum | null,
    updatedAt: string,
    updatedBy: string,
    valueOfAcquisition?: number | null,
    valueOfClick?: number | null,
    valueOfPost?: number | null,
    writeGroups?: Array< string | null > | null,
  } | null,
};

export type GetCampaignPostQueryVariables = {
  campaignPostId: string,
};

export type GetCampaignPostQuery = {
  getCampaignPost?:  {
    __typename: "CampaignPost",
    acquisitions?:  {
      __typename: "ModelAcquisitionConnection",
      nextToken?: string | null,
    } | null,
    campaign?:  {
      __typename: "Campaign",
      campaignId: string,
      clientAccountId: string,
      confirmationPageUrl?: string | null,
      createdAt: string,
      createdBy: string,
      description?: string | null,
      disabled?: boolean | null,
      eventId: string,
      modalIncentiveTCs?: string | null,
      modalIncentiveText?: string | null,
      modalPattern?: string | null,
      modalPatternLocationEnum?: CampaignModalPatternLocationEnum | null,
      modalText?: string | null,
      modalTitle?: string | null,
      name?: string | null,
      postForwardUrl?: string | null,
      postImageUrl?: string | null,
      postText?: string | null,
      readGroups?: Array< string | null > | null,
      registrationPageUrl?: string | null,
      targetInfluencerTypeEnum?: CampaignTargetInfluencerTypeEnum | null,
      updatedAt: string,
      updatedBy: string,
      valueOfAcquisition?: number | null,
      valueOfClick?: number | null,
      valueOfPost?: number | null,
      writeGroups?: Array< string | null > | null,
    } | null,
    campaignId: string,
    campaignPostClicks?:  {
      __typename: "ModelCampaignPostClickConnection",
      nextToken?: string | null,
    } | null,
    campaignPostId: string,
    clientAccountId: string,
    country?: string | null,
    createdAt: string,
    createdBy: string,
    disabled?: boolean | null,
    email?: string | null,
    emailVerified?: string | null,
    familyName?: string | null,
    givenName?: string | null,
    language?: string | null,
    linkedInPostId?: string | null,
    name?: string | null,
    pictureUrl?: string | null,
    postType?: string | null,
    postTypeEnum?: CampaignPostPostTypeEnum | null,
    readGroups?: Array< string | null > | null,
    sub?: string | null,
    updatedAt: string,
    updatedBy: string,
    writeGroups?: Array< string | null > | null,
  } | null,
};

export type GetCampaignPostClickQueryVariables = {
  campaignPostClickId: string,
};

export type GetCampaignPostClickQuery = {
  getCampaignPostClick?:  {
    __typename: "CampaignPostClick",
    campaign?:  {
      __typename: "Campaign",
      campaignId: string,
      clientAccountId: string,
      confirmationPageUrl?: string | null,
      createdAt: string,
      createdBy: string,
      description?: string | null,
      disabled?: boolean | null,
      eventId: string,
      modalIncentiveTCs?: string | null,
      modalIncentiveText?: string | null,
      modalPattern?: string | null,
      modalPatternLocationEnum?: CampaignModalPatternLocationEnum | null,
      modalText?: string | null,
      modalTitle?: string | null,
      name?: string | null,
      postForwardUrl?: string | null,
      postImageUrl?: string | null,
      postText?: string | null,
      readGroups?: Array< string | null > | null,
      registrationPageUrl?: string | null,
      targetInfluencerTypeEnum?: CampaignTargetInfluencerTypeEnum | null,
      updatedAt: string,
      updatedBy: string,
      valueOfAcquisition?: number | null,
      valueOfClick?: number | null,
      valueOfPost?: number | null,
      writeGroups?: Array< string | null > | null,
    } | null,
    campaignId: string,
    campaignPost?:  {
      __typename: "CampaignPost",
      campaignId: string,
      campaignPostId: string,
      clientAccountId: string,
      country?: string | null,
      createdAt: string,
      createdBy: string,
      disabled?: boolean | null,
      email?: string | null,
      emailVerified?: string | null,
      familyName?: string | null,
      givenName?: string | null,
      language?: string | null,
      linkedInPostId?: string | null,
      name?: string | null,
      pictureUrl?: string | null,
      postType?: string | null,
      postTypeEnum?: CampaignPostPostTypeEnum | null,
      readGroups?: Array< string | null > | null,
      sub?: string | null,
      updatedAt: string,
      updatedBy: string,
      writeGroups?: Array< string | null > | null,
    } | null,
    campaignPostClickId: string,
    campaignPostId: string,
    createdAt: string,
    createdBy: string,
    disabled?: boolean | null,
    postForwardUrl?: string | null,
    readGroups?: Array< string | null > | null,
    updatedAt: string,
    updatedBy: string,
    writeGroups?: Array< string | null > | null,
  } | null,
};

export type GetCampaignPostStateQueryVariables = {
  campaignPostStateId: string,
};

export type GetCampaignPostStateQuery = {
  getCampaignPostState?:  {
    __typename: "CampaignPostState",
    campaign?:  {
      __typename: "Campaign",
      campaignId: string,
      clientAccountId: string,
      confirmationPageUrl?: string | null,
      createdAt: string,
      createdBy: string,
      description?: string | null,
      disabled?: boolean | null,
      eventId: string,
      modalIncentiveTCs?: string | null,
      modalIncentiveText?: string | null,
      modalPattern?: string | null,
      modalPatternLocationEnum?: CampaignModalPatternLocationEnum | null,
      modalText?: string | null,
      modalTitle?: string | null,
      name?: string | null,
      postForwardUrl?: string | null,
      postImageUrl?: string | null,
      postText?: string | null,
      readGroups?: Array< string | null > | null,
      registrationPageUrl?: string | null,
      targetInfluencerTypeEnum?: CampaignTargetInfluencerTypeEnum | null,
      updatedAt: string,
      updatedBy: string,
      valueOfAcquisition?: number | null,
      valueOfClick?: number | null,
      valueOfPost?: number | null,
      writeGroups?: Array< string | null > | null,
    } | null,
    campaignId: string,
    campaignPostStateId: string,
    createdAt: string,
    createdBy: string,
    disabled?: boolean | null,
    state?: string | null,
    updatedAt: string,
    updatedBy: string,
  } | null,
};

export type GetClientAccountQueryVariables = {
  clientAccountId: string,
};

export type GetClientAccountQuery = {
  getClientAccount?:  {
    __typename: "ClientAccount",
    address1?: string | null,
    address2?: string | null,
    annualEventCount: string,
    averageEventAttendance: string,
    city?: string | null,
    clientAccountId: string,
    companyName: string,
    country?: string | null,
    createdAt: string,
    createdBy: string,
    currencyCode: string,
    disabled?: boolean | null,
    events?:  {
      __typename: "ModelEventConnection",
      nextToken?: string | null,
    } | null,
    industry: string,
    licenseKey?: string | null,
    licenseKeyValidUntil?: string | null,
    pixelInstallationTrackings?:  {
      __typename: "ModelPixelInstallationTrackingConnection",
      nextToken?: string | null,
    } | null,
    readGroups?: Array< string | null > | null,
    state?: string | null,
    updatedAt: string,
    updatedBy: string,
    users?:  {
      __typename: "ModelUserConnection",
      nextToken?: string | null,
    } | null,
    writeGroups?: Array< string | null > | null,
    zip?: string | null,
  } | null,
};

export type GetErrorLogQueryVariables = {
  errorLogId: string,
};

export type GetErrorLogQuery = {
  getErrorLog?:  {
    __typename: "ErrorLog",
    awsRequestId?: string | null,
    campaignId?: string | null,
    clientAccountId?: string | null,
    createdAt: string,
    createdBy: string,
    disabled?: boolean | null,
    errorLogId: string,
    errorMessage?: string | null,
    functionName?: string | null,
    metadata?: string | null,
    readGroups?: Array< string | null > | null,
    stack?: string | null,
    updatedAt: string,
    updatedBy: string,
    writeGroups?: Array< string | null > | null,
  } | null,
};

export type GetEventQueryVariables = {
  eventId: string,
};
// ...
export type GetPixelInstallationTrackingQueryVariables = {
  pixelInstallationTrackingId: string,
};
// ...
export type GetUserQueryVariables = {
  userId: string,
};
// ...
export type ListAcquisitionsQueryVariables = {
  acquisitionId?: string | null,
  filter?: ModelAcquisitionFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};
// ...
export type ListCampaignPostClicksQueryVariables = {
  campaignPostClickId?: string | null,
  filter?: ModelCampaignPostClickFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};